signv4: new AWS Signv4 API to perform HTTP signatures
The following patch introduces a new API interface to add signatures
to HTTP requests. The interface is based in the official Signv4
spec by AWS:

  - https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html

The purpose of this interface is to be used by plugins that requires to
access AWS services with Signv4 signature headers, the patch exposes one
specific function:

  flb_sds_t flb_signv4_do(struct flb_http_client *c, int normalize_uri,
                          int amz_date,
                          time_t t_now,
                          char *access_key,
                          char *region, char *service,
                          char *secret_key, char *security_token);

When creating a plugin that connects to an AWS service, once all headers
and credentials are set, you can sign the flb_http_client context with
the function above, the code will take care of the following internal tasks:

 - Task 1: Create a Canonical Request for Signature Version 4
 - Task 2: Create a String to Sign for Signature Version 4
 - Task 3: Calculate the Signature for AWS Signature Version 4
 - Task 4: Add the Signature to the HTTP Request

If the function succeed, it will return a flb_sds_t string with the signature
value, note that the signature header will be appended automatically to the
flb_http_client context, the caller is responsible to destroy the flb_sds_t
returned context. If the function fails it returns NULL.

== Notes ==

- The API is exposed only when at build time the CMake flag FLB_SIGNV4 is
  enabled (default is enabled).

- There are not external dependencies, all the implementation is done through the
  usage of the core Fluent Bit API.

- The implementation has passed 100% of the AWS Signv4 test suite. In the next
  patch I am adding the internal unit tests for validation:

  https://docs.aws.amazon.com/general/latest/gr/signature-v4-test-suite.html

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>